feat(backend): attach request's id to logs + extract log logic
